Created user_commands cog file that currently implements alive and whoami. Also discovered how to restrict commands to role ids in the user_commands file, I think this can be implemented per command easily Could possibly apply to an entire cog file, but not sure yet. It would still make sense to separate commands by user level, such as user_commands
